Making a fresh new file system
Django app inside Typing_practice_app and front end files inside Gaurav's frontend file.
