+++ .gitignore
@@ -1,0 +1,1 @@
+Typing_practice_app/static/Symbiosis/javascript/typing.js
