+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\ASUS\\Anaconda3\\envs\\WebDjango\\python.exe"
+}
