+++ Gaurav's front end files/index.html
@@ -1,0 +1,50 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Typing Practice</title>
+    <link rel="stylesheet" href="styling.css">
+    <script src="interface.js" defer></script>
+    <script src="typing practice.js" defer></script>
+</head>
+
+<body>
+    <nav id="navbar">
+        <ul>
+            <li id="sitename">Typing Practice</li>
+            <li id="about" class="navbtn">About Us</li>
+            <li id="goback" class="navbtn">Go Back</li>
+        </ul>
+    </nav>
+    <section id="parentcontainer1">
+        <div id="container" class="same">
+            <div id="txtcont"><textarea name="txt">are</textarea></div>
+            <div id='lettercont'></div>
+
+        </div>
+    </section>
+
+    <section id="parentcontainer2">
+        <div id="display"><button id="restartbtn" class="utilitybtn">Restart &#10227</button>
+            <button id="nxtpara"  class="utilitybtn">Next Paragraph</button>
+        </div>
+    </section>
+
+    <section id="aboutus">
+        <div class="same">
+            <div id="aboutcont">
+
+            </div>
+
+        </div>
+
+    </section>
+
+
+
+
+</body>
+
+</html>
