+++ Typing_practice_app/Symbiosis/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class SymbiosisConfig(AppConfig):
+    name = 'Symbiosis'
