+++ Typing_practice_app/Symbiosis/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('', views.rendertypingPage, name="rendertypingPage"),
+]
