--- styling.css
@@ -1,208 +1,0 @@
-*{
-    margin: 0;
-    padding: 0;
-   
-}
-nav{
-    background-color: rgb(155, 109, 197);
-    padding: 8px;
-}
-
-ul{display: flex;
-    list-style: none;
-    align-items: center;
-    
-}
-
-#sitename{
-    font-size: 3rem;
-    position: relative;
-    left:8%
-}
-
-.navbtn{
-    cursor: pointer;
-    padding: 1px 8px;
-    font-size: 1.2rem;
-    position: absolute;
-    right:10%;
-    border: 2px white solid;
-    border-radius: 11px;
-    background-color: rgb(175, 207, 187);
-}
-#goback{
-    display: none;
-}
-
-#parentcontainer1{
-    margin : 2% 18%;
-    margin-top: 8%;
-    padding: 8px;
-
-    display: flex;
-    /* height: 40vh; */
-    justify-content: center;
-    align-items: center;
-}
-#parentcontainer2{
-    margin :0% 18%;
-    margin-bottom: 2%;
-    
-    padding: 8px;
-
-    display: flex;
-    height: 12vh;
-    
-    align-items: center;
-}
-
-
-.same{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 8px;
-    border: 2px rgba(146, 139, 146, 0.692) solid;
-    padding: 8px 20px;
-    width: 100%;
-    min-height: 15vh;
-    
-    
-}
-.selected{
-    border: 1px rgb(0, 119, 255) solid;
-}
-
-.wrong{
-     
-    background-color:  rgba(255, 0, 0, 0.308);
-}
-
-.right{
-    color: rgb(0, 119, 255);
-}
-
-span{
-    padding: 0px;
-    margin: 0px;
-    font-family: 'Source Code Pro', monospace, Courier;
-    font-size: 1.1rem;
-    line-height: 2em;
-    letter-spacing: 0.09em;
-    min-width: 2px;
-}
-
-.space{
-    padding:0px 5px;
-    margin: 0px;
-    font-family: 'Source Code Pro', monospace, Courier;
-    font-size: 1.1rem;
-    line-height: 2em;
-    letter-spacing: 0.09em;
-    min-width: 2px;
-}
-textarea{
-    z-index: 0;
-    
-    /* left: 1%; */
-    resize: none;
-    width: 100%; 
-    height: 100%;
-    
-    opacity: 0;
-}
-
-.utilitybtn{
-    font-size: 1.05rem;
-    margin: 2% 1%;
-    margin-left: 5%;
-
-    
-    padding: 2px 8px;
-    border: rgb(112, 167, 212) solid 2px;
-    border-radius: 8px;
-    
-    transition: 0.1;
-    outline: none;
-    
-}
-
-.utilitybtn:hover{
-    
-    
-    padding: 2px 8px;
-    
-    color: rgb(255, 255, 255);
-    background-color:rgb(112, 167, 212);
-    border-radius: 8px;
-    /* transform: scale(1.05); */
-    cursor: pointer;
-}
-
-    
-
-
-
-
-#result{
-    margin :0% 16%;
-    
-    padding: 8px;
-    padding-top: 20px;
-
-    display: flex;
-    font-size: 1.4rem;
-    justify-content: space-evenly;
-    align-items: center;
-
-}
-#txtcont{
-    position: absolute;
-    display: flexbox;
-    width: 60%;
-    
-    min-height: 15vh;
-    
-
-}
-
-#lettercont{
-    
-    width: 100%;
-    min-height: 15vh;
-    
-}
-
-#aboutus{
-    display: none;
-    margin : 2% 18%;
-    margin-top: 8%;
-    padding: 8px;
-
-    /* display: flex; */
-    /* height: 40vh; */
-    justify-content: center;
-    align-items: center;
-    
-}
-
-#aboutcont{
-    width: 100%;
-    min-height: 15vh;
-    height: 15vh;
-    
-    
-}
-
-#display{
-    display: flex;
-    
-    justify-content: flex-end;
-    align-items: center; 
-    
-    padding: 8px 20px;
-    width: 100%;
-    min-height: 15vh;
-    height: 15vh;
-
-}
